requirements.txt
gradio>=3.0
rembg>=2.0
pillow>=9.0
onnxruntime>=1.18